+++ apps/show/src/routes/_layout.programme.$day/route-params.ts
@@ -1,0 +1,6 @@
+import { ShowDay } from "#core/show-day";
+import { zu } from "@animeaux/zod-utils";
+
+export const routeParamsSchema = zu.object({
+  day: zu.nativeEnum(ShowDay.Enum),
+});
